+++ lib/authorizer/aws4.js
@@ -1,7 +1,11 @@
 var aws4 = require('aws4'),
+    fs = require('fs'),
     _ = require('lodash'),
-    util = require('../runner/util'),
-    RequestBody = require('postman-collection').RequestBody;
+    crypto = require('crypto'),
+    qs = require('qs'),
+    RequestBody = require('postman-collection').RequestBody
+
+    BODY_HASH_HEADER = 'X-Amz-Content-Sha256';
 
 /**
  * @implements {AuthHandlerInterface}
@@ -93,6 +97,45 @@ module.exports = {
         return aws4.sign(params, params.credentials);
     },
 
+    computeBodyHash: function (body, algorithm, digestEncoding, callback) {
+        if (!body || !algorithm) { return callback(); }
+
+        var hash = crypto.createHash(algorithm);
+
+        if (body.mode === RequestBody.MODES.raw) {
+            hash.update(body.toString());
+
+            return callback(hash.digest(digestEncoding));
+        }
+
+        if (body.mode === RequestBody.MODES.urlencoded) {
+            hash.update(qs.stringify(_.transform(body.urlencoded.all(), function (result, query) {
+                result[query.key] = query.value;
+            }, {})));
+
+            return callback(hash.digest(digestEncoding));
+        }
+
+        if (body.mode === RequestBody.MODES.formdata) {
+            // @todo: This requires adding new data to form instead of setting headers for AWS auth
+
+            return callback();
+        }
+
+        if (body.mode === RequestBody.MODES.file) {
+            // eslint-disable-next-line security/detect-non-literal-fs-filename
+            var stream = fs.createReadStream(body.file.src);
+
+            stream.on('data', function (chunk) {
+                hash.update(chunk);
+            });
+
+            stream.on('end', function () {
+                callback(hash.digest(digestEncoding));
+            });
+        }
+    },
+
     /**
      * Signs a request.
      *
@@ -127,9 +170,9 @@ module.exports = {
             });
         }
 
-        util.bodyHash(request.body, 'sha256', function(bodyHash) {
+        self.computeBodyHash(request.body, 'sha256', 'hex', function(bodyHash) {
             bodyHash && request.upsertHeader({
-                key: 'X-Amz-Content-Sha256',
+                key: BODY_HASH_HEADER,
                 value: bodyHash,
                 system: true
             });
