+++ lib/runner/util.js
@@ -1,10 +1,4 @@
-var fs = require('fs'),
-    _ = require('lodash'),
-    crypto = require('crypto'),
-    sdk = require('postman-collection'),
-    qs = require('qs'),
-
-    /**
+var /**
      * @const
      * @type {string}
      */
@@ -120,49 +114,5 @@ module.exports = {
             // eslint-disable-next-line security/detect-non-literal-fs-filename
             callback(null, resolver.createReadStream(fileSrc));
         });
-    },
-
-    bodyHash: function(body, algorithm, callback) {
-        if (!body) { return callback(); }
-
-        var hash = crypto.createHash(algorithm);
-
-        if (body.mode === sdk.RequestBody.MODES.file) {
-            // eslint-disable-next-line security/detect-non-literal-fs-filename
-            this.readFile(body.file.src, function (chunk) {
-                hash.update(chunk);
-            }, function () {
-                callback(hash.digest('hex'));
-            });
-
-            return;
-        }
-
-        if (body.mode === sdk.RequestBody.MODES.formdata) {
-            // @todo: This requires adding new data to form instead of setting headers for AWS auth
-            callback();
-
-            return;
-        }
-
-        if (body.mode === sdk.RequestBody.MODES.urlencoded) {
-            hash.update(qs.stringify(_.transform(body.urlencoded.all(), function (result, query) {
-                result[query.key] = query.value;
-            }, {})));
-
-            return callback(hash.digest('hex'));
-        }
-
-        hash.update(body.toString());
-
-        return callback(hash.digest('hex'));
-    },
-
-    readFile: function (src, cbData, cbEnd) {
-        // eslint-disable-next-line security/detect-non-literal-fs-filename
-        var stream = fs.createReadStream(src);
-
-        stream.on('data', cbData);
-        stream.on('end', cbEnd);
     }
 };
